Google search scraper

+++ server-side/model/page_data_model.go
@@ -6,6 +6,6 @@ type PageData struct {
 	ID             uint `gorm:"primarykey"`
 	CreatedAt      time.Time
 	UpdatedAt      time.Time
-	SearchResultID uint64
+	SearchResultID uint
 	HtmlData       string `gorm:"type:text"`
 }
